+++ packages/sdk-react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.4.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-react-v0.3.0...@unidy.io/sdk-react-v0.4.0) (2026-07-01)
+
+
+### Features
+
+* [UD-3063] sdk-react invited step + resend invitation ([#297](https://github.com/UnidyID/unidy-sdk/issues/297)) ([3bbf87b](https://github.com/UnidyID/unidy-sdk/commit/3bbf87bfebe684c93d978f33e73ac7b7ed49fa31))
+
+## [0.3.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-react-v0.2.0...@unidy.io/sdk-react-v0.3.0) (2026-05-13)
+
+
+### Features
+
+* [UD-2524] wire ServicesService into React SDK ([#272](https://github.com/UnidyID/unidy-sdk/issues/272)) ([a4e08de](https://github.com/UnidyID/unidy-sdk/commit/a4e08de205fb81b2bfa30f4cb84d96f75011a311))
+
 ## [0.2.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-react-v0.1.0...@unidy.io/sdk-react-v0.2.0) (2026-05-07)