+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,63 @@
 # Changelog
 
+## [1.9.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.8.0...@unidy.io/sdk-v1.9.0) (2026-07-14)
+
+
+### Features
+
+* [UD-3099] SDK invitation acceptance ([#308](https://github.com/UnidyID/unidy-sdk/issues/308)) ([9f82a08](https://github.com/UnidyID/unidy-sdk/commit/9f82a0813b78eb2d1f29582ab71c1ffaaa03f022))
+
+
+### Bug Fixes
+
+* add slot="empty" support to ticketable-list and transaction-list ([#303](https://github.com/UnidyID/unidy-sdk/issues/303)) ([797f4c6](https://github.com/UnidyID/unidy-sdk/commit/797f4c6f3a8b9baa71cc4ddd479c5c919b9fb705))
+* pagination controls broken without manually wired store  ([#302](https://github.com/UnidyID/unidy-sdk/issues/302)) ([708887a](https://github.com/UnidyID/unidy-sdk/commit/708887a6040a0685dd75d9de0438c648a298b667))
+
+## [1.8.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.7.1...@unidy.io/sdk-v1.8.0) (2026-07-01)
+
+
+### Features
+
+* [UD-3063] SDK invited step and resend invitation flow ([#296](https://github.com/UnidyID/unidy-sdk/issues/296)) ([a74dfc1](https://github.com/UnidyID/unidy-sdk/commit/a74dfc1bf6aa9160ac1eb75e17b4c977b9d141a3))
+* [UD-3063] sdk-react invited step + resend invitation ([#297](https://github.com/UnidyID/unidy-sdk/issues/297)) ([3bbf87b](https://github.com/UnidyID/unidy-sdk/commit/3bbf87bfebe684c93d978f33e73ac7b7ed49fa31))
+* [UD-3064] add Norwegian as a supported language ([#298](https://github.com/UnidyID/unidy-sdk/issues/298)) ([368e912](https://github.com/UnidyID/unidy-sdk/commit/368e912dd069d88682235945cd63f5bbff10909e))
+
+## [1.7.1](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.7.0...@unidy.io/sdk-v1.7.1) (2026-06-26)
+
+
+### Bug Fixes
+
+* [UD-2217] clear pending recovery step on reset-password redirect ([#290](https://github.com/UnidyID/unidy-sdk/issues/290)) ([544e1f1](https://github.com/UnidyID/unidy-sdk/commit/544e1f15ef328462574c71e07b7dabe9d1feea05))
+
+## [1.7.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.6.0...@unidy.io/sdk-v1.7.0) (2026-06-02)
+
+
+### Features
+
+* [UD-2721] expose global-logout prop on u-logout-button ([#284](https://github.com/UnidyID/unidy-sdk/issues/284)) ([f1501b8](https://github.com/UnidyID/unidy-sdk/commit/f1501b89968f802e8e124d027b4fa5e474f97dfb))
+
+
+### Bug Fixes
+
+* [UD-2550] remove window.location.href default from standalone createSignIn ([#274](https://github.com/UnidyID/unidy-sdk/issues/274)) ([f9e7161](https://github.com/UnidyID/unidy-sdk/commit/f9e71619d3654e2da42e5fa6e236b57bc0b723a8))
+* expose missing transaction fields from API response ([#280](https://github.com/UnidyID/unidy-sdk/issues/280)) ([3cd75e9](https://github.com/UnidyID/unidy-sdk/commit/3cd75e92fc2ea5667d07eb4261e03df08002e6f0))
+* social login button sentry noise ([#279](https://github.com/UnidyID/unidy-sdk/issues/279)) ([0284157](https://github.com/UnidyID/unidy-sdk/commit/0284157e158173863080663e5636c93a58a0635f))
+
+## [1.6.0](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.5.1...@unidy.io/sdk-v1.6.0) (2026-05-13)
+
+
+### Features
+
+* [UD-2524] add ServicesService with list and disconnect endpoints ([#271](https://github.com/UnidyID/unidy-sdk/issues/271)) ([6dac731](https://github.com/UnidyID/unidy-sdk/commit/6dac7319d42df6c573c7122bc1f4ca23266a0d79))
+
+
+### Bug Fixes
+
+* [UD-2147] handle already_subscribed without auto-sending login email ([#268](https://github.com/UnidyID/unidy-sdk/issues/268)) ([ad829c3](https://github.com/UnidyID/unidy-sdk/commit/ad829c3ab876843af4b31f1e17196d5ac15daa26))
+* [UD-2374] SDK email language respects UI language ([#267](https://github.com/UnidyID/unidy-sdk/issues/267)) ([8628b4d](https://github.com/UnidyID/unidy-sdk/commit/8628b4deaccd0d00d603ca8a49d5ecf7b58741a0))
+* [UD-2475] align transaction line_items schema with API response ([#265](https://github.com/UnidyID/unidy-sdk/issues/265)) ([a16aa0f](https://github.com/UnidyID/unidy-sdk/commit/a16aa0f55e2e2cead8d8064e985a3daa5aa88953))
+* [UD-2531] tolerate any jsonb shape in ticketable metadata/wallet_export ([#270](https://github.com/UnidyID/unidy-sdk/issues/270)) ([969f292](https://github.com/UnidyID/unidy-sdk/commit/969f2924f84d6e8a3bf2652360da20df31f6d23f))
+
 ## [1.5.1](https://github.com/UnidyID/unidy-sdk/compare/@unidy.io/sdk-v1.5.0...@unidy.io/sdk-v1.5.1) (2026-05-05)